// repositorio/SIMTR/4-implementacao/simtr-web/src/app/cliente/consulta-cliente/aba-nivel-documental/model/loading-model-nivel-documental.ts
export class LoadingModelNivelDocumental {
    documentacaoValida: boolean = true;
    countDocumentacaoValida: number = 0;
    countTotalDocumentoValidacao: number = 0;

    documentacaoPendente: boolean = true;
    countDocumentacaoPendente: number = 0;
    countTotalDocumentoPendente: number = 0;
            
    documentacaoTipoDocumento: boolean = true;
    countDocumentacaoTipoDocumento: number = 0;
    countTotalDocumentoTipoDocumento: number = 0;
}